Move constants module to api folder
